Returned old onjection. Moved sorting of posts to data-layer method. Returned router-link instead of html-link in episode list compoent.
