+++ shared/backend/get_blogs.js
@@ -1,5 +1,8 @@
 import each from 'lodash/each';
 import filter from 'lodash/filter';
+import sort from 'lodash/sortBy';
+
+import moment from 'moment';
 
 const ignoredKeys = ['latest', 'guid'];
 function isIgnoredKey(key) {
@@ -37,6 +40,14 @@ function matchingLimit(blog, index, limit) {
     return (index < limit);
 }
 
+function compare(dateTimeA, dateTimeB) {
+    const momentA = moment(dateTimeA,"YYYY-MM-DD");
+    const momentB = moment(dateTimeB,"YYYY-MM-DD");
+    if (momentA > momentB) return 1;
+    else if (momentA < momentB) return -1;
+    else return 0;
+}
+
 module.exports = {
     get_blogs: function (req, res, blogmetadata_map, env, exclude = ['/episodes', '/transcripts']) {
         env = (env === 'prod') ? 'master' : 'dev';
@@ -58,12 +69,16 @@ module.exports = {
 
         // remove unnecessary keys
         let blogs = filter(blogmetadata_map, (blog, id) => !isIgnoredKey(id));
-
         // filter only relative blogs
         blogs = filter(blogs, (blog, id) => isMatchingQuery(blog, {url, exclude, env}));
 
         // calculate total matched blogs count
         const total = blogs.length;
+
+        blogs = blogs.sort((a, b) => {
+            // custom sort by publish data
+            return compare(b['publish_date'], a['publish_date']);
+        });
 
         // slice over limits
         blogs = blogs
